added Pixelization object which is contained for mappers and the like
