+++ autoarray/__init__.py
@@ -19,8 +19,8 @@ from .fit.fit_imaging import FitImaging
 from .fit.fit_interferometer import FitInterferometer
 from .instruments import acs
 from .instruments import euclid
-from .inversion.mappers.abstract import AbstractMapper
-from .inversion import mesh as pix
+from .inversion.pixelization.mappers.abstract import AbstractMapper
+from .inversion.pixelization import mesh
 from .inversion import regularization as reg
 from .inversion.inversion.settings import SettingsInversion
 from .inversion.inversion.abstract import AbstractInversion
@@ -32,13 +32,13 @@ from .inversion.inversion.factory import (
 from .inversion.inversion.factory import (
     inversion_interferometer_unpacked_from as InversionInterferometer,
 )
-from .inversion.mappers.factory import mapper_from as Mapper
-from .inversion.mappers.rectangular import MapperRectangularNoInterp
-from .inversion.mappers.delaunay import MapperDelaunay
-from .inversion.mappers.voronoi import MapperVoronoiNoInterp
-from .inversion.mappers.voronoi import MapperVoronoi
-from .inversion.mesh.abstract import AbstractMesh
-from .inversion.mesh.settings import SettingsPixelization
+from .inversion.pixelization.mappers.factory import mapper_from as Mapper
+from .inversion.pixelization.mappers.rectangular import MapperRectangularNoInterp
+from .inversion.pixelization.mappers.delaunay import MapperDelaunay
+from .inversion.pixelization.mappers.voronoi import MapperVoronoiNoInterp
+from .inversion.pixelization.mappers.voronoi import MapperVoronoi
+from .inversion.pixelization.mesh.abstract import AbstractMesh
+from .inversion.pixelization.settings import SettingsPixelization
 from .inversion.inversion.imaging.mapping import InversionImagingMapping
 from .inversion.inversion.imaging.w_tilde import InversionImagingWTilde
 from .inversion.inversion.interferometer.w_tilde import InversionInterferometerWTilde
