+++ autoarray/inversion/inversion/interferometer/w_tilde.py
@@ -9,7 +9,7 @@ from autoarray.inversion.inversion.interferometer.abstract import (
 from autoarray.dataset.interferometer import WTildeInterferometer
 from autoarray.inversion.linear_obj.linear_obj import LinearObj
 from autoarray.inversion.inversion.settings import SettingsInversion
-from autoarray.inversion.mappers.abstract import AbstractMapper
+from autoarray.inversion.pixelization.mappers.abstract import AbstractMapper
 from autoarray.inversion.regularization.abstract import AbstractRegularization
 from autoarray.preloads import Preloads
 from autoarray.operators.transformer import TransformerNUFFT
