+++ autoarray/inversion/pixelization/mappers/rectangular.py
@@ -3,8 +3,8 @@ from typing import Dict, Optional, Tuple
 
 from autoconf import cached_property
 
-from autoarray.inversion.mappers.abstract import AbstractMapper
-from autoarray.inversion.mappers.abstract import PixSubWeights
+from autoarray.inversion.pixelization.mappers.abstract import AbstractMapper
+from autoarray.inversion.pixelization.mappers.abstract import PixSubWeights
 from autoarray.structures.arrays.uniform_2d import Array2D
 from autoarray.structures.grids.uniform_2d import Grid2D
