+++ autoarray/inversion/pixelization/mappers/voronoi.py
@@ -3,14 +3,14 @@ from typing import Dict, Optional, Tuple
 
 from autoconf import cached_property
 
-from autoarray.inversion.mappers.abstract import AbstractMapper
-from autoarray.inversion.mappers.abstract import PixSubWeights
+from autoarray.inversion.pixelization.mappers.abstract import AbstractMapper
+from autoarray.inversion.pixelization.mappers.abstract import PixSubWeights
 from autoarray.structures.arrays.uniform_2d import Array2D
 from autoarray.structures.grids.uniform_2d import Grid2D
 from autoarray.structures.grids.sparse_2d import Grid2DSparse
 
 from autoarray.numba_util import profile_func
-from autoarray.inversion.mappers import mapper_util
+from autoarray.inversion.pixelization.mappers import mapper_util
 
 
 class AbstractMapperVoronoi(AbstractMapper):
