+++ autoarray/inversion/pixelization/mesh/voronoi.py
@@ -6,10 +6,10 @@ from autoarray.structures.grids.uniform_2d import Grid2D
 from autoarray.structures.grids.sparse_2d import Grid2DSparse
 from autoarray.structures.mesh.voronoi_2d import Mesh2DVoronoi
 from autoarray.preloads import Preloads
-from autoarray.inversion.mesh.abstract import AbstractMesh
-from autoarray.inversion.mesh.settings import SettingsPixelization
-from autoarray.inversion.mappers.voronoi import MapperVoronoiNoInterp
-from autoarray.inversion.mappers.voronoi import MapperVoronoi
+from autoarray.inversion.pixelization.mesh.abstract import AbstractMesh
+from autoarray.inversion.pixelization.settings import SettingsPixelization
+from autoarray.inversion.pixelization.mappers.voronoi import MapperVoronoiNoInterp
+from autoarray.inversion.pixelization.mappers.voronoi import MapperVoronoi
 
 from autoarray.numba_util import profile_func
