+++ autoarray/inversion/plot/inversion_plotters.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from autoconf import conf
 
-from autoarray.inversion.mappers.abstract import AbstractMapper
+from autoarray.inversion.pixelization.mappers.abstract import AbstractMapper
 from autoarray.plot.abstract_plotters import Plotter
 from autoarray.plot.mat_wrap.visuals import Visuals2D
 from autoarray.plot.mat_wrap.include import Include2D
