+++ autoarray/structures/mesh/rectangular_2d.py
@@ -10,7 +10,7 @@ from autoarray.structures.arrays.uniform_2d import Array2D
 from autoarray.structures.mesh.abstract_2d import Abstract2DMesh
 
 from autoarray.structures.grids import grid_2d_util
-from autoarray.inversion.mesh import mesh_util
+from autoarray.inversion.pixelization.mesh import mesh_util
 from autoarray import type as ty
