+++ test_autoarray/inversion/pixelization/mappers/test_rectangular.py
