+++ test_autoarray/inversion/pixelization/mesh/test_abstract.py
@@ -3,7 +3,7 @@ import autoarray as aa
 
 class TestRegression:
     def test__grid_is_relocated_via_border(self, sub_grid_2d_7x7):
-        pixelization = aa.pix.VoronoiMagnification(shape=(3, 3))
+        pixelization = aa.mesh.VoronoiMagnification(shape=(3, 3))
 
         mask = aa.Mask2D.circular(
             shape_native=(60, 60),
@@ -41,7 +41,7 @@ class TestRegression:
         assert sparse_grid[0, 0] != mapper.source_mesh_grid[0, 0]
         assert mapper.source_mesh_grid[0, 0] < 5.0
 
-        pixelization = aa.pix.VoronoiNNMagnification(shape=(3, 3))
+        pixelization = aa.mesh.VoronoiNNMagnification(shape=(3, 3))
 
         mapper = pixelization.mapper_from(
             source_grid_slim=grid,
